refactor: update controller
