+++ README.md
@@ -7,19 +7,19 @@
 - [x] 사용자로부터 자동차의 이름들을 한 줄에 입력받는다.
     - [x] 입력의 시작이나 끝이 공백으로만 이루어져 있다면 제거한다.
     
+- [ ] 구분자를 기준으로 각각의 자동차를 생성한다.
     - #### 예외 처리
-    - [x] 입력이 없는 경우 (개행 문자만 입력된 경우)
-    - [x] 구분자로만 입력된 경우 (자동차의 수가 0이므로)
-
-- [ ] 구분자를 기준으로 각의 자동차 이름을 검사한다. (자동차의 이름은 5자 이하다.)
-
-    - #### 예외 처리
-    - [ ] 자동차 이름이 5자를 초과하면 에러 문구를 띄운다.
-
+    - [x] 차를 저장한 리스트의 길이가 0인 경우
+    - [ ] 각각의 자동차 이름이 5자를 초과하는 경우
+    
 - [ ] 사용자로부터 이동 횟수를 입력받는다.
+    - #### 예외처리
+    - [ ] 숫자가 아닌 경우
+    
 - [ ] 턴 마다 각 자동차의 랜덤값을 구한다.
     - [ ] 랜덤값이 4이상인 자동차만 한 칸 전진한다.
     - [ ] 해당 턴의 진행 상황을 출력한다.
+    
 - [ ] 자동차들중 가장 많이 전진한 자동차(들)의 이름을 출력한다.  
 
 <br>
