+++ src/main/java/racingcar/Car.java
@@ -1,7 +1,5 @@
 package racingcar;
 
-import error.customException.NotValidInputException;
-
 public class Car {
     private final String name;
     private int position = 0;
@@ -13,6 +11,10 @@ public class Car {
     // 추가 기능 구현
     public void moveCarForward () {
         position ++;
+    }
+
+    public String getName () {
+        return name;
     }
 
 }
